Fix tests setup
