+++ __mocks__/WebView.js
@@ -1,0 +1,3 @@
+export default function WebView() {
+  return null;
+}
