+++ jest.config.js
@@ -1,3 +1,6 @@
 module.exports = {
   preset: 'react-native',
+  moduleNameMapper: {
+    'react-native-webview': '<rootDir>/__mocks__/WebView.js',
+  },
 };
